Fix #135324: Data-block selector icon drawn too big
DPI scaling was applied twice. Added an API comment to make this more
clear for the caller of the function. This is exactly the changes to
main committed in 47bbf3fcaefb2c63661a8ab9098f840925c67a9f, which is a
correction to 1f88645728.

Pull Request: https://projects.blender.org/blender/blender/pulls/135411